Municipalities with classes

Joined together data set with administrative areas, calculated population loss/gain category and geometry

Usage

taantuvat

Format

Data frame with columns

kunta_nro

Three number municipality code

kunta_nimi

Municipality name

kokmuutos_2010_2022

Total population change from 2010 to 2022

suht_muutos_2010_2022

Population change from 2010 to 2022 relative to population

luokka

six categories used to describe level of population loss/gain

NUTS 1

Tukialue

ELY-keskus

Suuralue

Maakunta

Seutukunta

Leader

Alueluokka

Kuntaryhma

geom

geometry column

Source

`kunnat`, `aluejaot`, `vaestonmuutos`.

Examples

if (require("sf")) {

taantuvat
}
#> Simple feature collection with 309 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 83747.59 ymin: 6637032 xmax: 732907.7 ymax: 7776431
#> Projected CRS: ETRS89 / TM35FIN(E,N)
#> First 10 features:
#>    kunta_nro kunta_nimi kokmuutos_2010_2022 suht_muutos_2010_2022
#> 1        005   Alajärvi               -1390            -14.074195
#> 2        009  Alavieska                -329            -12.509140
#> 3        010     Alavus               -1484            -12.543318
#> 4        016   Asikkala                -537             -6.494799
#> 5        018     Askola                 -68             -1.373801
#> 6        019       Aura                 125              3.151301
#> 7        020       Akaa                -385             -2.290104
#> 8        035     Brändö                 -48            -10.351692
#> 9        043     Eckerö                  15              1.583563
#> 10       046  Enonkoski                -276            -18.966064
#>                   luokka           NUTS 1     Tukialue
#> 1  Voimakkaasti taantuva     Manner-Suomi tukialue 2+3
#> 2  Voimakkaasti taantuva     Manner-Suomi   tukialue 1
#> 3  Voimakkaasti taantuva     Manner-Suomi tukialue 2+3
#> 4               Taantuva     Manner-Suomi tukialue 2+3
#> 5        Hieman taantuva     Manner-Suomi tukialue 2+3
#> 6         Hieman kasvava     Manner-Suomi tukialue 2+3
#> 7        Hieman taantuva     Manner-Suomi tukialue 2+3
#> 8  Voimakkaasti taantuva Landskapet Åland        Åland
#> 9         Hieman kasvava Landskapet Åland        Åland
#> 10 Voimakkaasti taantuva     Manner-Suomi   tukialue 1
#>                   ELY-keskus                    Suuralue
#> 1    11 Etelä-Pohjanmaan ely           3 Länsi-Suomen sa
#> 2  13 Pohjois-Pohjanmaan ely 4 Pohjois- ja Itä-Suomen sa
#> 3    11 Etelä-Pohjanmaan ely           3 Länsi-Suomen sa
#> 4              04 Hämeen ely           2 Etelä-Suomen sa
#> 5           01 Uudenmaan ely     1 Helsinki-Uudenmaan sa
#> 6    02 Varsinais-Suomen ely           2 Etelä-Suomen sa
#> 7          05 Pirkanmaan ely           3 Länsi-Suomen sa
#> 8                   16 Åland         5 Ålands storområde
#> 9                   16 Åland         5 Ålands storområde
#> 10        07 Etelä-Savon ely 4 Pohjois- ja Itä-Suomen sa
#>                    Maakunta              Seutukunta
#> 1    14 Etelä-Pohjanmaan mk      146 Järviseudun sk
#> 2  17 Pohjois-Pohjanmaan mk       177 Ylivieskan sk
#> 3    14 Etelä-Pohjanmaan mk    144 Kuusiokuntien sk
#> 4       07 Päijät-Hämeen mk           071 Lahden sk
#> 5           01 Uudenmaan mk          015 Porvoon sk
#> 6    02 Varsinais-Suomen mk          025 Loimaan sk
#> 7          06 Pirkanmaan mk 063 Etelä-Pirkanmaan sk
#> 8       21 Landskapet Åland     213 Ålands skärgård
#> 9       21 Landskapet Åland    212 Ålands landsbygd
#> 10        10 Etelä-Savon mk      103 Savonlinnan sk
#>                                       Leader               Alueluokka
#> 1                             40 Aisapari ry             Ydinmaaseutu
#> 2                        44 Rieska-LEADER ry             Ydinmaaseutu
#> 3                           37 Kuudestaan ry             Ydinmaaseutu
#> 4                      15 Päijänne-Leader ry Kaupunkien läh. maaseutu
#> 5   01 Maaseudun kehittämisyhdistys SILMU ry Kaupunkien läh. maaseutu
#> 6  06 Varsinais-Suomen jokivarsikumppanit ry Kaupunkien läh. maaseutu
#> 7                         16 Pirkan Helmi ry Kaupunkien läh. maaseutu
#> 8              88 Ahvenanmaan ei-Leader-alue  Harvaan asuttu maaseutu
#> 9              88 Ahvenanmaan ei-Leader-alue             Ydinmaaseutu
#> 10                         26 Piällysmies ry  Harvaan asuttu maaseutu
#>               Kuntaryhma                           geom
#> 1  Taajaan asutut kunnat MULTIPOLYGON (((366787.9 70...
#> 2  Maaseutumaiset kunnat MULTIPOLYGON (((382543.4 71...
#> 3  Taajaan asutut kunnat MULTIPOLYGON (((343298.2 69...
#> 4  Taajaan asutut kunnat MULTIPOLYGON (((436139.7 67...
#> 5  Maaseutumaiset kunnat MULTIPOLYGON (((426631 6720...
#> 6  Maaseutumaiset kunnat MULTIPOLYGON (((263938.3 67...
#> 7  Taajaan asutut kunnat MULTIPOLYGON (((328844.1 67...
#> 8  Maaseutumaiset kunnat MULTIPOLYGON (((176190.4 67...
#> 9  Maaseutumaiset kunnat MULTIPOLYGON (((92735.28 67...
#> 10 Maaseutumaiset kunnat MULTIPOLYGON (((600317.4 69...